Create VoyagerUserController and moved all logic there simplified a lot changes
