+++ publishable/database/seeds/DataTypesTableSeeder.php
@@ -19,7 +19,7 @@ class DataTypesTableSeeder extends Seeder
                 'icon'                  => 'voyager-person',
                 'model_name'            => 'TCG\\Voyager\\Models\\User',
                 'policy_name'           => 'TCG\\Voyager\\Policies\\UserPolicy',
-                'controller'            => '',
+                'controller'            => '\TCG\Voyager\Http\Controllers\VoyagerUserController',
                 'generate_permissions'  => 1,
                 'description'           => '',
             ])->save();
