+++ resources/views/profile.blade.php
@@ -20,6 +20,6 @@
         <h4>{{ ucwords(Auth::user()->name) }}</h4>
         <div class="user-email text-muted">{{ ucwords(Auth::user()->email) }}</div>
         <p>{{ Auth::user()->bio }}</p>
-        <a href="{{ route('voyager.users.editProfile', Auth::user()->id) }}" class="btn btn-primary">{{ __('voyager::profile.edit') }}</a>
+        <a href="{{ route('voyager.users.edit', Auth::user()->id) }}" class="btn btn-primary">{{ __('voyager::profile.edit') }}</a>
     </div>
 @stop
