+++ routes/voyager.php
@@ -32,7 +32,7 @@ Route::group(['as' => 'voyager.'], function () {
         Route::post('logout', ['uses' => $namespacePrefix.'VoyagerController@logout',  'as' => 'logout']);
         Route::post('upload', ['uses' => $namespacePrefix.'VoyagerController@upload',  'as' => 'upload']);
 
-        Route::get('profile', ['uses' => $namespacePrefix.'VoyagerController@profile', 'as' => 'profile']);
+        Route::get('profile', ['uses' => $namespacePrefix.'VoyagerUserController@profile', 'as' => 'profile']);
 
         try {
             foreach (DataType::all() as $dataType) {
@@ -127,16 +127,6 @@ Route::group(['as' => 'voyager.'], function () {
         ], function () use ($namespacePrefix) {
             Route::get('/', ['uses' => $namespacePrefix.'VoyagerCompassController@index',  'as' => 'index']);
             Route::post('/', ['uses' => $namespacePrefix.'VoyagerCompassController@index',  'as' => 'post']);
-        });
-
-        //Profile
-        Route::group([
-            'as'     => 'users.',
-            'prefix' => 'profile',
-        ], function () use ($namespacePrefix) {
-            Route::get('/{id}', ['uses' => $namespacePrefix.'VoyagerController@showProfile', 'as' => 'showProfile']);
-            Route::get('/{id}/edit', ['uses' => $namespacePrefix.'VoyagerController@editProfile', 'as' => 'editProfile']);
-            Route::put('/{id}', ['uses' => $namespacePrefix.'VoyagerController@updateProfile', 'as' => 'updateProfile']);
         });
 
         event(new RoutingAdminAfter());
